main: more sophisticated ssh_wait
I was still seeing occasional blips in which ssh worked fine during the
ssh_wait procedure, and then proceeded to die one or two seconds later
when e.g. syncing the repo or sending the worker script.

Change things so that we require a sustained 5 seconds connectivity
before accepting the node as properly initialized.