spawner.py: check if testsuite should run before spawning
Previously, we checked in the testrunner if the branch for which we were
triggered was among the branches defined in the YAML. It's more
efficient to do this check directly in the spawner, so that we can avoid
any spawning at all if necessary.

Now, we can be sure that the testrunner is always running because a test
needs to run. This also simplifies the assumption made in
count_failures(). That is, if we get an rc 0 from the testrunner but yet
don't have an rc file (meaning no test was actually run), we can assume
that something was semantically wrong with the YAML file itself.