testrunner: use error if provisioning goes wrong
Let's reserve the 'failure' state for tests returning nonzero. If
anything happens during provisioning that fails, even if it may be due
to a mistyped value in the YAML, let's return 'error'. We will maintain
this property for the upcoming 'prepare' key as well.

The idea is that we should 'error' on things that users don't even
*expect* to fail, and fall outside the actual testing procedure.