spawner.py: upload a basic results summary page (#24)
* spawner.py: upload a basic results summary page

Setting the commit status URL to the branch itself when there are
testsuites marked required is not very intuitive. Additionally, the
commit can disappear from the branch, giving users no way to discover
the statuses.

We now instead link to a very simple summary page containing the links
to all the other test results. When redhat-ci is used in combination
with homu, this will be accessible directly from the PR through the
comment homu posts.

Closes: #23