spec: also set required context on PRs
For the same reasons the 'required' context is useful for branches, it
can also be useful for PRs: it allows repo admins to set a single status
as required in the branch protection settings rather than multiple,
possibly changing, statuses.

This also specifically helps with status-based exemptions in Homu: by
having the required context set on the PR, Homu will be able to
determine whether it can safely skip re-testing the code on the auto
branch.

Closes: #30
Approved by: cgwalters