spec: add full support for UTF-8
The YAML file is now required to be in UTF-8. This allows users to have
non-ASCII ☃❤❄ characters in YAML comments. Additionally, non-ASCII
characters are also supported in two fields: as values to environment
variables and as part of the tests list. All other fields will yield a
ParserError if non-ASCII characters are specified.

We also improve error handling for PyYAML and pykwalify errors and
rethrow them as ParserError exceptions.