testrunner: fix checking whether we're on AH
Not being containerized means that we're either virtualized or
clustered. In the case of clustered, we may or may not be container or
host controlled. The correct thing to check here when we want to know if
we're on an Atomic Host is really whether if we're host-controlled
(which happens in the virtualized mode and can happen in clustered
mode), if that host is an Atomic Host. Fix the conditional
appropriately (which otherwise could have tried to "vmssh" even though
we're container controlled).

This should have been refactored during the introduction of the cluster
environment.