spec: drop support for nightly streams for now
These images are broken on the internal OpenStack instance. Let's drop
support for them for now, and circle back to doing it properly in the
future, which will require the images to work properly as well as
formalizing a config.yaml or site.yaml.

Also drop support for rhel/7/atomic and rhel/7/atomic/stage. This will
require some work in the backend so let's wait until we have a good
use case for this. All projects today should be able to use
centos/7/atomic anyway.