testrunner: inject private addr in hosts file
The openshift-ansible installer was complaining that on each cluster
node, their hostname wasn't resolving to an IP address of one of their
interfaces. Which is a pretty valid point actually.

We fix that here by ensuring that we filter out the public IP of the
host when inserting in its /etc/hosts.