codebase: rename to PAPR
This is the first step towards renaming the project to PAPR. We keep
things compatible by still checking for the old YAML file name as well
as injecting variables under the old name.

Rework the README.md a bit and split out the details of running it on
your own in a separate document.