testrunner: safer error handling
I was reading the cleanup logs when we encounter failing infra and found
out that we're still trying to tear down nodes even when they weren't
even provisioned at all. The core of the issue is that
`x=$(cat /non/existent)` will not trigger `errexit`, so we have to
explicitly check for the file beforehand. Accomodating these shell
gotchas is a short term solution until we absorb all the logic into the
Python bits.

Also make sure that the gh module writes its diagnostics to stderr
rather than stdout.