os provisioning: use volumes for secondary disk
The new OpenStack instance doesn't support flavors with ephemeral disks.
Rather, they encourage users to provision Cinder volumes. We teach our
provisioner to do this in this patch. This is required to support
projects like atomic and container-storage-setup which want a secondary
disk for testing.

I didn't bother cleaning up the code much. A lot of this is already
structured much more nicely in the Python rewrite and I didn't want to
duplicate more work than necessary.