testrunner: add support for PAPR_PULL_TARGET_BRANCH
It can be useful for pull request tests to know which branch they're
targeted at. E.g. openshift-ansible for example can make use of this in
order to match the version of origin to use.

Tests could have used the API themselves anonymously before this, though
given that the rate limit is 60 requests per hour per IP, that's not
really scalable. So let's just make the request ourselves and pass down
that information.

Closes: #78
Approved by: miabbott