ocp: wait for a pod to be scheduled
Part of the appeal of switch to OCP is the fact that pods are actually
queued and scheduled when there's resources for them. When a pod is
accepted but not yet scheduled due to resource constraints, its phase is
set to "Pending". Allow pods to be in the pending state for a while, but
error out if it's anything else.