utils/common: wait longer before SSH comes up
In kernel release 4.16.4, there was a change to how `/dev/random` is
handled at boot time. This is causing much longer startup times than
before. There's an issue open to track this[1], but in the meantime,
let's do something no one likes to do and increase our timeout before
deeming a node broken.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1572944