testrunner: don't use tee
Dump the output directly to the file instead of passing it through `tee`
first. My hypothesis is that `tee` can fail and artificially cause `rc`
to be set to `1` even though the test itself passed if ssh or the swarm
process is too slow to read the output.

What leads me to believe this is that looking at the `output.log` that
gets uploaded, one sees much more output than what actually ends up in
the console. So I suspect a bunch of data gets buffered but not actually
flushed in time until it backs up enough to cause a `write()` from `tee`
to return `EAGAIN`, at which point it dies.

There are some flags we could maybe use to control this, but anyway, we
don't really need the output to be sent to the Jenkins console as well.
It makes things noisy and we can just fetch it out of `output.log`.