os_provision: add random sleep interval before provisioning
I have a strong suspicion that part of the reason we're hitting a lot of
OpenStack failures is because it can't handle all the simultaneous
requests. Right now in e.g. rpm-ostree, we're essentially spawning 7
hosts at the same time. And what I've noticed is that the very first
ones we spawn often never become reachable over SSH.

Haven't filed a ticket yet, but let's try this as well in case it works
around the issue until it's fixed.